+++ src/AgeSimulation02.java
@@ -2,6 +2,7 @@ import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Scanner;
+import Model.*;
 
 public class AgeSimulation02 {
     public static void main(String[] args) {
@@ -34,69 +35,9 @@ public class AgeSimulation02 {
         }
     }
 
-    public static class User {
-        private String name;
-        private LocalDate birth;
 
-        public User(String name, LocalDate birth) {
-            this.name = name;
-            this.birth = birth;
-        }
 
-        public String getName() {
-            return name;
-        }
 
-        public LocalDate getBirth() {
-            return birth;
-        }
-    }
 
-    static class UserRepository {
-        private static int sequence = 0;
-        private static Map<Integer, User> users = new HashMap<>();
 
-        public UserRepository() {
-            users.put(++sequence, new User("홍길동", LocalDate.of(1990, 1, 1)));
-            users.put(++sequence,
-                new User("박한수", LocalDate.of(2000, LocalDate.now().getMonth(), LocalDate.now().getDayOfMonth())));
-            users.put(++sequence, new User("강호동", LocalDate.of(2005, 6, 1)));
-        }
-
-        public User save(User user) {
-            users.put(++sequence, user);
-            return user;
-        }
-
-        public User findById(int id) {
-            return users.get(id);
-        }
-    }
-
-    public static class AgeCalculator {
-
-        public boolean isUnderAge(LocalDate birth) {
-            int age = getAge(birth);
-            return age < 20;
-        }
-
-        public int getAge(LocalDate birth) {
-            LocalDate now = LocalDate.now();
-
-            int age = now.getYear() - birth.getYear() + 1;
-            // age += isOverBirthDay(now, birth) ? 1 : 0;
-            return age;
-        }
-
-        // private boolean isOverBirthDay(LocalDate now, LocalDate birth) {
-        //     if (now.getMonthValue() != birth.getMonthValue())
-        //         return now.getMonthValue() > birth.getMonthValue();
-        //     return now.getDayOfMonth() > birth.getDayOfMonth();
-        // }
-
-        public boolean isBirthDay(LocalDate birth) {
-            LocalDate now = LocalDate.now();
-            return now.getMonthValue() == birth.getMonthValue() && now.getDayOfMonth() == birth.getDayOfMonth();
-        }
-    }
 }
