+++ src/View/InputView.java
@@ -1,0 +1,4 @@
+package View;
+
+public class InputView {
+}
